requirements.txt
aiohttp
PyYAML
requests
beautifulsoup4
rank-bm25
